Use A2W_XOSC0 register instead of delays
I'm not quite sure how this works, but I imagine that reading from
this register blocks until the next oscillator tick. The clock edge
from the oscillator is in fact exactly what the OTP memory chip
needs for some changes to take effect, so this approach should be
better in all circumstances than hardcoding a delay (which depends
on the VPU clock).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>