Add a delay in uart_init
My CP2102 regulary receives garbage on the serial port. It seems
that the last symbol is not fully transmitted and my UART does not
find the start of the next frame correctly.

Keeping the TX signal low for two more bits makes the receiver
abort the current frame and sync again with the next one.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>